+++ source/extensions/isaacsim.core.includes/docs/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+
+## [2.5.4] - 2025-12-06
+### Changed
+- Run clang tidy
+- Migrate BaseResetNode to Events 2.0.
+
 ## [2.5.3] - 2025-11-07
 ### Changed
 - Update to Kit 109 and Python 3.12