+++ source/extensions/isaacsim.core.nodes/docs/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+
+## [5.1.3] - 2025-12-06
+### Changed
+- Migrate OgnOnPhysicsStep to Events 2.0.
+- Update description
+
 ## [5.1.2] - 2025-12-05
 ### Changed
 - Migrate to Events 2.0.