+++ source/extensions/isaacsim.robot.manipulators.examples/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.2.4] - 2025-12-07
+### Changed
+- Updated description
+
 ## [1.2.3] - 2025-11-27
 ### Changed
 - Add missing docstrings